api/requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-jose[cryptography]
passlib[bcrypt]

sqlalchemy
alembic
psycopg2
python-dotenv

torch==2.1.2+cu121
torchaudio==2.1.2+cu121
torchvision==0.16.2+cu121

boto3
pytest